Move merge and select symbols to the top level
These are core component to work with channels, so they are moved to the
`frequenz.channels` top level module.
