+++ src/frequenz/channels/_anycast.py
@@ -38,9 +38,9 @@ class Anycast(Generic[_T]):
     thread-safe.
 
     When there are multiple channel receivers, they can be awaited
-    simultaneously using [select][frequenz.channels.util.select],
-    [Merge][frequenz.channels.util.Merge] or
-    [MergeNamed][frequenz.channels.util.MergeNamed].
+    simultaneously using [select][frequenz.channels.select],
+    [Merge][frequenz.channels.Merge] or
+    [MergeNamed][frequenz.channels.MergeNamed].
 
     Example:
         ``` python
