+++ src/frequenz/channels/_broadcast.py
@@ -32,9 +32,9 @@ class Broadcast(Generic[_T]):
     are thread-safe.  Because of this, `Broadcast` channels are thread-safe.
 
     When there are multiple channel receivers, they can be awaited
-    simultaneously using [select][frequenz.channels.util.select],
-    [Merge][frequenz.channels.util.Merge] or
-    [MergeNamed][frequenz.channels.util.MergeNamed].
+    simultaneously using [select][frequenz.channels.select],
+    [Merge][frequenz.channels.Merge] or
+    [MergeNamed][frequenz.channels.MergeNamed].
 
     Example:
         ``` python
