+++ src/frequenz/channels/_merge_named.py
@@ -1,58 +1,41 @@
 # License: MIT
 # Copyright © 2022 Frequenz Energy-as-a-Service GmbH
 
-"""Merge messages coming from channels into a single stream."""
-
+"""Merge messages coming from channels into a single stream containing name of message."""
 import asyncio
 import itertools
 from collections import deque
 from typing import Any, TypeVar
 
-from .._receiver import Receiver, ReceiverStoppedError
+from ._receiver import Receiver, ReceiverStoppedError
 
 _T = TypeVar("_T")
 
 
-class Merge(Receiver[_T]):
-    """Merge messages coming from multiple channels into a single stream.
+class MergeNamed(Receiver[tuple[str, _T]]):
+    """Merge messages coming from multiple named channels into a single stream.
 
-    Example:
-        For example, if there are two channel receivers with the same type,
-        they can be awaited together, and their results merged into a single
-        stream, by using `Merge` like this:
-
-        ```python
-        from frequenz.channels import Broadcast
-
-        channel1 = Broadcast[int](name="input-chan-1")
-        channel2 = Broadcast[int](name="input-chan-2")
-        receiver1 = channel1.new_receiver()
-        receiver2 = channel2.new_receiver()
-
-        merge = Merge(receiver1, receiver2)
-        while msg := await merge.receive():
-            # do something with msg
-            pass
-        ```
-
-        When `merge` is no longer needed, then it should be stopped using
-        `self.stop()` method. This will cleanup any internal pending async tasks.
+    When `MergeNamed` is no longer needed, then it should be stopped using
+    `self.stop()` method. This will cleanup any internal pending async tasks.
     """
 
-    def __init__(self, *args: Receiver[_T]) -> None:
-        """Create a `Merge` instance.
+    def __init__(self, **kwargs: Receiver[_T]) -> None:
+        """Create a `MergeNamed` instance.
 
         Args:
-            *args: sequence of channel receivers.
+            **kwargs: sequence of channel receivers.
         """
-        self._receivers: dict[str, Receiver[_T]] = {
-            str(id): recv for id, recv in enumerate(args)
-        }
+        self._receivers: dict[str, Receiver[_T]] = kwargs
+        """The sequence of channel receivers to get the messages to merge."""
+
         self._pending: set[asyncio.Task[Any]] = {
             asyncio.create_task(anext(recv), name=name)
             for name, recv in self._receivers.items()
         }
-        self._results: deque[_T] = deque(maxlen=len(self._receivers))
+        """The set of pending tasks to merge messages."""
+
+        self._results: deque[tuple[str, _T]] = deque(maxlen=len(self._receivers))
+        """The internal buffer of merged messages."""
 
     def __del__(self) -> None:
         """Cleanup any pending tasks."""
@@ -61,7 +44,7 @@ class Merge(Receiver[_T]):
                 task.cancel()
 
     async def stop(self) -> None:
-        """Stop the `Merge` instance and cleanup any pending tasks."""
+        """Stop the `MergeNamed` instance and cleanup any pending tasks."""
         for task in self._pending:
             task.cancel()
         await asyncio.gather(*self._pending, return_exceptions=True)
@@ -98,16 +81,17 @@ class Merge(Receiver[_T]):
                 if isinstance(item.exception(), StopAsyncIteration):
                     continue
                 result = item.result()
-                self._results.append(result)
+                self._results.append((name, result))
                 self._pending.add(
+                    # pylint: disable=unnecessary-dunder-call
                     asyncio.create_task(anext(self._receivers[name]), name=name)
                 )
 
-    def consume(self) -> _T:
+    def consume(self) -> tuple[str, _T]:
         """Return the latest value once `ready` is complete.
 
         Returns:
-            The next value that was received.
+            The next key, value that was received.
 
         Raises:
             ReceiverStoppedError: if the receiver stopped producing messages.
@@ -123,10 +107,10 @@ class Merge(Receiver[_T]):
     def __str__(self) -> str:
         """Return a string representation of this receiver."""
         if len(self._receivers) > 3:
-            receivers = [str(p) for p in itertools.islice(self._receivers.values(), 3)]
+            receivers = [str(p) for p in itertools.islice(self._receivers, 3)]
             receivers.append("…")
         else:
-            receivers = [str(p) for p in self._receivers.values()]
+            receivers = [str(p) for p in self._receivers]
         return f"{type(self).__name__}:{','.join(receivers)}"
 
     def __repr__(self) -> str:
